// src/states/model/ovp.model.ts
export interface OvpModel {
    id?: number;
    empCode?: string;
    sinNum?: string;
    ovpDatcrea?: string;
    ovpDatsigne?: string;
    ovpDatdeb?: string;
    ovpDatfin?: string;
    ovpNbVirm?: number;
    ovpUserCode?: string;
    ovpMont?: string;
    op?: number;
    banquePayeur?: string;
    agence?: string;
    cpteOrdonnateur?: string;
    banqueCreditee?: string;
    cpteCredite?: string;
    posteCptable?: string;
    delegation?: string;
    ovpDateCtl?: string;
    ovpChargCode?: string;
    ovpMontCrean?: number;
    motifDate?: string;
    motifUser?: string;
    reference?: string;
    ovpModifDate?: string;
    matric?: string;
    ovpTraite?: string;
}